+++ src/Links/Changelog.md
@@ -7,7 +7,7 @@
 
 - Added the unidirectional links feature
 - Added the possibility to retrieve the link items directly using `linkItems()` method as `Get::the($type)->linkItems()->of($model)`
-- Added the `link_items` helper (shortcut to Get::the()->linkItems()
+- Added the `link_items` helper, shortcut to Get::the()->linkItems()
 
 ## 4.0.0
 ##### 2024-04-25